requirements.txt
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
pandas>=2.0
openpyxl>=3.1
chardet>=5.2
Pillow>=10.1
